Only check the AuthnRequest destination if the request is signed
To check if the request is signed, we check if the current policy is to
ignore the signature. In all other case if we reach the
check_destination(), it means that signature validation passed.

In the future Lasso should be modified to report the signature
validation status in profile.signatureStatus whatever the signature
verification hint. Currently this is done when processing responses but
not when processing requests (ref #1577).