Allow frontends to handle completely their content
If a frontend has a method like:

		def login(self, request, *args, **kwargs):
			...

then normal handling through the form() and post() method is ignored and
this method is used instead. If it returns an HTTP response with a
status code of 200 the content is extracted and used in the login page.

If any other response is returned, it's taken as the login page result,
which allows a frontend to short-circuit the normal login workflow.