ldap_backend: reimplement password change
Now it uses the password change extended
operation, it can be deactivated by setting 'use_password_modify' to False
or if 'active_directory' is True.

On AD if the old password is not passed to the modify_password() function,
and LDAP replace is used instead of a delete followed by an add.