idp/saml: collapse attribute values (fixes #6378)
Coninuation of commit 6ceeff9.

Previous patch was too simple, this one remove any lasso specialization
in the SAMLAttribute model. The method to_lasso_attribute was renamed
to_tuples, it now returns tuples maed of the name, the name format, a
friendly name or None and a value. Lasso specific are now completely in
the add_attributes() function of the saml2_endpoints module.

Tuples are aggregated arround their lasso.Attribute target in a
dictionnary indexed by name and name_format. The dictionnary is
initialized using attributes already present in the assertion, for retro
compatibility with attribute aggregator until its removal. Only values
having an unique MiscTextNode with textChild equals to true are
considered.