custom_user/models: restore default value of User.last_login (fixes #7525)
Removing them broke the password reset form. Currently moving the code
base to Django 1.8 cannot be done without breaking compatibility with
pre-Django 1.8 as we depend on the abstract model AbstractUser which set
the default values for the last_login field. We will have to duplicate
AbstractUser when making the jump to Django 1.8.