a2_rbac: delete permissions first in OrganizationalUnit.delete() (fixes #9504)
To prevent constraint check failures on deletion Django update all nullable
fields before deleting dependant models in the CASCADE behaviour instead of
computing a dependency graph. But it interracts badly with the partial index
created on Permission model fields, which enforce unicity even if field
Permission.ou, when you try to delete an organizational unit, it tracebacks an
integrity constraint error because it cannot update the ou field of admin role
of the organizational unit being deleted.