compatibility fix to use of add_to_blocks with Django > 1.8 (fixes #10436)
Starting with Django 1.8 context processors are loaded until the effective
rendering of a template, and their state is immediately destroyed after it; for
addtoblock/renderblock template tags it means they will not work anymore when
rendering of different templates is not done in only one rendering phase, as
with the login view. In this case we must explicitely initialize the
add_to_blocks variable in the context_instance, which will override and outlive
the default variable created by the authentic2 context processor.