registration: add comment to explain redirect on POST
If we detected some users with the same email and the email_is_unique flag is
on, we must redirect back to the same view without the 'create' parameter as
creation of a new user is forbidden.

It can happend if there is a race condition between to submit to the same form
(in different tabs or because the user clicked multiple times).