crypto: add deterministic encryption (fixes #17537)
Data is encrypted using a key, a fixed salt and hash of the data.

Salt is hashed with a chosen hash algorithm. The same hash algorithm is used to
hash datas.

The key is improved using PBKDF2. by default with a count of 1 and same salt
used for generating the IV for encryption.

IV is generated by xoring hash of the salt with hash of the data.

Transmitted hash is shortened to key size of the used algorithm (here AES128).

Count, name of the hash algorithm, base64url encoded value of the hash of the
data and base64url encoding of the ciphered text are joined by a dollar
character.

Deterministic encryption is useful to generate reversible, targeted, opaque to
third parties pseudonymous such as OpenID Connect sub or SAML NameID or
SessionIndex.